Fix beams in malformed MusicXML files if option 'fix_beams' is enabled
    * ImportOptions class renamed MusicXmlOptions. Some changes for
      removing Builder class and to add public setters.
    * Added option 'fix_beams' to MusicXmlOptions class. By default it
      is enabled. MusicXmlOptions accesible from LomseDoorway and from
      LibraryScope objects.
    * MusicXML importer modified for detecting badformed beams and fixing
      them when option 'fix_beams' is enabled.
    * Code refactored for moving 'Autobeamer' to its own file. Class
      LmdAutoBeamer removed.
